+++ docs/about/release-notes.rst
@@ -6,7 +6,7 @@ Release Notes
 
 .. Template, copy this to create a new section after a release:
 
-   v23.01.0
+   vRELEASE_PLACEHOLDER
    --------
 
    Features
@@ -34,6 +34,42 @@ Release Notes
    Neil Vaytet :sup:`a`\ ,
    Jan-Lukas Wynen :sup:`a`\ ,
    and Sunyoung Yoo :sup:`a`
+
+
+vRELEASE_PLACEHOLDER
+--------
+
+Features
+~~~~~~~~
+
+* Added new string-formatting options `#3017 <https://github.com/scipp/scipp/pull/3017>`_.
+* Add ``size`` property to :class:`scipp.Variable` and :class:`scipp.DataArray`.
+
+Breaking changes
+~~~~~~~~~~~~~~~~
+
+Bugfixes
+~~~~~~~~
+
+* Fix access to value of 0-D variable with "spatial transform" dtypes using the ``value`` property.
+  Previously this raised an exception if the dtype was ``affine_transform3``, ``translation3``, or ``rotation3`` `#3033 <https://github.com/scipp/scipp/pull/3033>`_.
+
+Documentation
+~~~~~~~~~~~~~
+
+Deprecations
+~~~~~~~~~~~~
+
+Stability, Maintainability, and Testing
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+Contributors
+~~~~~~~~~~~~
+
+Simon Heybrock :sup:`a`\ ,
+Neil Vaytet :sup:`a`\ ,
+Jan-Lukas Wynen :sup:`a`\ ,
+and Sunyoung Yoo :sup:`a`
 
 
 v23.01.0